added UpgradeCard() to CardManager
takes a name and a stat. finds the card to upgrade from the cards queue and upgrades the selected stat of the card. mana cannot be reduced below 1. damage increase is infinite. returns true if an upgrade occurred, returns false if stat is max, no card called name exists, or no stat called stat exists.